Move handling of Constants from Midgard to TN
