+++ src/renderer/services/thorchain/index.ts
@@ -23,6 +23,8 @@ const {
   reloadThornodeUrl,
   getNodeInfos$,
   reloadNodeInfos,
+  reloadThorchainConstants,
+  thorchainConstantsState$,
   inboundAddressesShared$,
   loadInboundAddresses$,
   reloadInboundAddresses,
@@ -73,6 +75,8 @@ export {
   interact$,
   getNodeInfos$,
   reloadNodeInfos,
+  reloadThorchainConstants,
+  thorchainConstantsState$,
   mimir$,
   reloadMimir,
   getLiquidityProviders,
